+++ CHANGELOG.md
@@ -6,12 +6,17 @@
 * pu-4 and wu-4 does not take effects.
 * reports who uploads now(SSE)
 * flag skip-validation
-* sdmin can see his/her answers/comments
+* admin can see his/her answers/comments
 * restrict comments to themselves
 * test deployment to eq.local
 * require reading comments before submit answers - what to do the first answerer?
   no comments available. not the first answerer only. before his first answer/comment,
   he has not yet sent any comments.
+
+
+# 0.3.14-SNAPSHOT
+
+* when no answers (so also no comments), no way to solve restrictions.
 
 # 0.3.13 (2025-09-25)